# server/openapi.yaml
openapi: 3.0.0
info:
  title: Communicator API 2026
  description: >
    API REST + WebSocket para sistema de mensajería instantánea y
    llamadas peer-to-peer (WebRTC). Soporta chat 1-a-1, grupos,
    presencia en línea, y señalización de llamadas de voz/video.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor Local de Desarrollo (ADB reverse)
  - url: http://192.168.XX.XX:3000
    description: Red WiFi local (reemplazar XX.XX con IP del servidor)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en /api/auth/login o /api/auth/register
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Credenciales inválidas
...
    AdminUser:
      type: object
      required: [id, username, email, avatar_url, is_online, last_seen, is_suspended, created_at]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        avatar_url:
          type: string
          nullable: true
        is_online:
          type: boolean
        last_seen:
          type: string
          format: date-time
          nullable: true
        invited_by:
          type: integer
          nullable: true
        invited_by_username:
          type: string
          nullable: true
        invited_at:
          type: string
          format: date-time
          nullable: true
        is_suspended:
          type: boolean
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Token JWT requerido o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadRequest:
      description: Solicitud mal formada o datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/admin/users:
    get:
      summary: Listar usuarios del workspace
      description: |
        Lista los usuarios del workspace activo del usuario autenticado
        (o del workspace indicado por query), con paginación.
        Solo miembros del workspace.
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: workspace_id
          in: query
          schema:
            type: integer
            description: 'ID del workspace a listar (default: workspace activo)'
      responses:
        '200':
          description: Lista paginada de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: No eres miembro del workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/users/{id}:
    patch:
      summary: Suspender/activar a un usuario
      description: |
        Solo un miembro del workspace. No puede modificarse a sí mismo.
        Cambio posible:
        - isSuspended (bool): suspender o reactivar la cuenta
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isSuspended:
                  type: boolean
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  updated:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El usuario no pertenece al workspace o no tienes permisos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Eliminar usuario del workspace
      description: |
        Solo un miembro del workspace. No puede eliminarse a sí mismo.
        Se elimina la membership; si no le quedan workspaces, se borra el usuario
        (borrado físico con ON DELETE CASCADE en messages/group_members).
      tags: [Admin]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario eliminado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  deleted:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: El usuario no pertenece al workspace o no tienes permisos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════════════════════
  # DEBUG / TEST (push notifications)
  # ══════════════════════════════════════════════
...